make core simple
