+++ hedgehog-lab/src/compiler.worker.js
@@ -1,0 +1,8 @@
+import * as Comlink from 'comlink'
+import { transpile } from './core'
+
+const compilerWorker = {
+  compile: (e) => transpile(e)
+}
+
+Comlink.expose(compilerWorker)
