+++ hedgehog-lab/src/output.worker.js
@@ -1,0 +1,9 @@
+import * as Comlink from 'comlink'
+import { executeOutput } from './core'
+
+const outputWorker = {
+  output: async (e) => executeOutput(e)
+}
+
+Comlink.expose(outputWorker)
+
